contrib/pkg/awstagdeprovision: Move cobra code into the hiveutil command
The cobra code is part of the command-line interface, and removing it
reduces the dependency load on folks who are calling the library
package from Go.
